+++ azext_edge/tests/edge/adr/conftest.py
@@ -10,20 +10,6 @@ from typing import Optional
 
 from azext_edge.edge.util.id_tools import parse_resource_id
 from ...generators import generate_random_string, get_zeroed_subscription
-from ...helpers import run
-
-
-@pytest.fixture()
-def require_init(init_setup):
-    # get the custom location used for tests.
-    if not all([init_setup.get("instanceName"), init_setup.get("resourceGroup")]):
-        pytest.skip("Cannot run this test without knowing the instance information.")
-
-    cluster_result = run(
-        f"az iot ops show -n {init_setup['instanceName']} -g {init_setup['resourceGroup']} "
-    )
-    init_setup["customLocationId"] = cluster_result["extendedLocation"]["name"]
-    yield init_setup
 
 
 @pytest.fixture()
